# Use Golang stlib instead of custom methods