--- a/uaparser/os.go
+++ b/uaparser/os.go
@@ -9,37 +9,13 @@ type Os struct {
 }
 
 func (parser *osParser) Match(line string, os *Os) {
-	matches := parser.Reg.FindStringSubmatch(line)
+	matches := parser.Reg.FindStringSubmatchIndex(line)
 	if len(matches) > 0 {
-		groupCount := parser.Reg.NumSubexp()
-
-		if len(parser.OSReplacement) > 0 {
-			os.Family = singleMatchReplacement(parser.OSReplacement, matches, 1)
-		} else if groupCount >= 1 {
-			os.Family = matches[1]
-		}
-
-		if len(parser.V1Replacement) > 0 {
-			os.Major = singleMatchReplacement(parser.V1Replacement, matches, 2)
-		} else if groupCount >= 2 {
-			os.Major = matches[2]
-		}
-
-		if len(parser.V2Replacement) > 0 {
-			os.Minor = singleMatchReplacement(parser.V2Replacement, matches, 3)
-		} else if groupCount >= 3 {
-			os.Minor = matches[3]
-		}
-
-		if len(parser.V3Replacement) > 0 {
-			os.Patch = singleMatchReplacement(parser.V3Replacement, matches, 4)
-		} else if groupCount >= 4 {
-			os.Patch = matches[4]
-		}
-
-		if groupCount >= 5 {
-			os.PatchMinor = matches[5]
-		}
+		os.Family = string(parser.Reg.ExpandString(nil, parser.OSReplacement, line, matches))
+		os.Major = string(parser.Reg.ExpandString(nil, parser.V1Replacement, line, matches))
+		os.Minor = string(parser.Reg.ExpandString(nil, parser.V2Replacement, line, matches))
+		os.Patch = string(parser.Reg.ExpandString(nil, parser.V3Replacement, line, matches))
+		os.PatchMinor = string(parser.Reg.ExpandString(nil, parser.V4Replacement, line, matches))
 	}
 }