* All test passes
* No loss in perf:
```
benchmark             old ns/op     new ns/op     delta
BenchmarkParser-8     26863572      26993123      +0.48%
```